Quick script to show table of "8-bit" charsets, like Latin-1
Print a quick table of so called "high ASCII" characters, that is,
all bytes from 0x80 to 0xFF. This is useful for testing locales such
as Latin-1 which use a single 8-bit byte instead of Unicode.

Output is sized to fit on a 40x16 terminal (for Tandy-200).

On a terminal with an 8-bit locale, such as Latin-1:
	mk8bit.sh	# Show 0x80 to 0xFF
	mk8bit.sh -a	# Show 0x20 to 0xFF (adds 7-bit ASCII).
	mk8bit.sh -A	# Show 0x00 to 0xFF (also sends control characters).

On a Unicode capable terminal:
	mk8bit.sh -f l1		# Convert from Latin-1 to UTF-8.
	mk8bit.sh -a -f t200	# Convert from Tandy-200 to UTF-8.