Test any many-to-one mappings in a charmap file
Usage: many-to-one.awk <filename.charmap>

Look for many-to-one mappings in the charmap definition and then
test to see if they actually work in the gconv module.

For example, both U+00B7 and U+2022 may map to the same character in
an 8-bit encoding.

Many-to-one mappings were often called for when dealing with
single-byte encodings which only had space for 256 characters.
Certain characters performed double or triple duty. For example,
Latin-1's 'ß' character, which is officially the German S-sharp, was
often used in English speaking countries to mean a Greek Beta.

Note: Mapping in reverse, each 8-bit character goes to a unique UCS
character. There is probably an official way to do this, but for
now, hackerb9 has decided that only the first UCS character listed
in the charmap file will be used for reverse maps.

SIDENOTE: glibc automatically handles many reasonable
transliterations when the //translit option is specified at the end
of the characterset name. For example, if a characterset defines
U+007C ('|') but not U+00A6 ('¦'), then any time the program prints
A6 to the screen, 7C will be sent instead:

	$ printf $'\uA6' | iconv -t ruthyn//translit | xxd -p
	7c

However, those transliterations are not what this program exercises.