Add 'test-iconv-charmap' to test using charmap file directly
iconv can use a file like ruthyn.charmap without needing a gconv
module. This new target tests to see if it works by using pre-created
data files in the tests/testdata/ directory.

    1. tests/testdata/RUTHYN-HD44780..UTF8
       A file containing all characters in the Ruthyn characterset.
       Encoded in UTF-8.

    2. tests/testdata/RUTHYN-HD44780
       Same as above, but encoded in the Ruthyn characterset.

    3. tests/testdata/RUTHYN-HD44780-sample.u8
       A sample presentable to the user to demonstrate functionality.
       Encoded in UTF-8.

Note that file 1 can be transformed into file 2, but not vice-versa.
This is because multiple UCS characters can map to a single Ruthyn
encoding.

For example, both U+0413 and U+0490 are encoded as the byte 0xA1:

<U0413>		/xA1		% CAPITAL_H (Г)
<U0490>		/xA1		% CAPITAL_G (Ґ)